kafka: do not return null record sets for fetch request
librdkafka and sarama treat a -1 (null) bytes field for the record set
as an error. redpanda returns a null value when some other error
scenario occurs when reading from the partition, but librdkafka tests
the recordset size before handling the errors. since some errors (like
leadership changed) are normal, this is a problem.

Signed-off-by: Noah Watkins <noah@vectorized.io>