storage: ensure log reader is closed cleanly
A consumer is free at any moment to declare that it no longer wants to
receive batches. The record batch reader will at that point stop and
release the underlying reader impl. However, there are some reader impls
that need to be cleaned up. For instance, log_reader needs to make sure
that it closes its internal reader. Before this patch the log reader
tried to _guess_ when it would no longer be used. This patch makes that
contract explicit: the record batch reader will invoke a shutdown method
on the reader impl for this type of clean-up.

Signed-off-by: Noah Watkins <noah@vectorized.io>