storage: remove compaction index on sealed segments
a segment may be compacted, but not have an associated compaction
segment writer, such as when a segment is opened or after a segment is
rolled. when segments are deleted compaction index deletion should be
dependent on if the segment is compacted not if the writer exists. this
was causing segment clean-up to miss compaction indexes.

fixes: [ch1239]

Signed-off-by: Noah Watkins <noah@vectorized.io>