storage: move record encoding to model namespace
The record batch is going to be store its record data in an encoded
format in memory and materialize records lazily. Therefore it needs
access to encoding/decoding routines. These had lived in the storage
namespace, but in order to avoid a dependency on storage:: from model::,
those routines are being moved.

Signed-off-by: Noah Watkins <noah@vectorized.io>