model: avoid copying batch when materializing records
Many contexts that iterate over batch records use a const ref to the
batch. This means that we could not share the batch data and had to make
a copy. This patch uses the iobuf_parser_base interface for parsing data
out of a const-ref iobuf, and builds an independent copy of records when
materializing records during iteration over records in a batch.

Signed-off-by: Noah Watkins <noah@vectorized.io>