cache: remove free pool list
In principle the batch cache pool is nice because it avoids the needless
reallocation of record batch headers as batches are removed and
re-inserted into the cache. However in practice we don't have a good
mechanism for pruning the pool. In some instances we've seen a spike of
say 500K record batches in the cache and then load drops and the pool is
holding on to 500K small objects from the small object memory allocation
pool.

Signed-off-by: Noah Watkins <noah@vectorized.io>