iobuf: use seastar memory alignment for iobuf copying
Memory allocations below 16K are handled by a small object pool in
seastar which guarnatees worst case about 15-25% memory overhead for
unaligned objects.

Above the 16K threshold, a large object pool is used which has 4K page
allocation granularity and pages are organized into contiguous spans of
lengths increasing by a power of 2. As an example, an allocation of
16K+1 will obtain a 32K span (8 pages), reserve 5 pages, and release the
remaining 3 pages back into the the pool.  The memory lost is thus 4095
bytes. For ~16K size batches this overhead can add up quickly.

We make the observation that when copying iobuf for record batches into
the batch cache we can optimize these allocations. When allocating from
the large pool we select the lower power of 2 bucket. For example if we
were to allocate 32K+3000 we'd make one large 32K allocation and satsify
the remainder 3000 bytes from the small pool.

Signed-off-by: Noah Watkins <noah@vectorized.io>