rpk/systemd: Get unit properties directly instead of listing units
The ListUnitsByName method is not available on systemd v219, which
is installed by default on AWS Linux 2 VMs. This was preventing
the fstrim tuner from working.

The 2 unit properties needed to evaluate whether the fstrim
service and timer are loaded and started - LoadState &
ActiveState -, can be retrieved through GetUnitProperty, which is
available in older systemd versions.