storage: ensure batch creation has consistent header
In the next patch the batch ctor is enhanced with some basic consistency
assertions, one of which is that the size in the header matches the data
payload size. This, in addition to other checks we may add that are
enabled in debug builds, batch construction needs to be changed so that
all of the inputs are consistent at construction time.

Signed-off-by: Noah Watkins <noah@vectorized.io>