storage: ensure segment truncation is durable
Segment truncation occurs on batch boundaries generally in order to
remove a batch from raft visibility. If a truncation isn't durable and a
crash occurs then upon recovery batches above the truncation offset may
appear to be part of live segment range.

However, even if the truncation point is made durable then there are
still challenges. Before truncation returns, but after the new
truncation point is made durable, the last in-memory chunk is hydrated
from disk to reflect any partially filled page (i.e. truncation occured
in the middle of a page).

If after truncation some appends occur to the in-memory chunk such that
more than 1 page contains data that need to be flushed another problem
can occur: suppose that after truncation at point T the disk looks like:

  [ ... T {EOF}]

but the in-memory chunk that has been apppended to but not yet flushed
looks like this:

  [ ... T <new data> P0 | P1 <new data> ... ]

where the first page P0 has been hydrated to reflect the partially
truncated page plus some new data, and the next page in the chunk also
contains some data.

Now when this chunk is flushed if P1 lands and the system crashes before
P0 lands on disk then the file length may also have been updated but
what would read in the hole between T and P1? At best zeros, at worst
the original "good batch" data. This case is natural with asynchronous
I/O and we rely on checksums to check this case.

Signed-off-by: Noah Watkins <noah@vectorized.io>