storage: ensure fallocate is disjoint from commit offset
Following an unaligned truncation the offset range of a chunk may
overlap the fallocation offset. Since commit offset is not updated until
chunk is full and commit index is used to determine when to extend
the fallocation point, a chunk may fill up and be dispatched and the
next append triggers and fallocation with an fallocation point below the
commit index, resulting in zeros overwriting a portion of the latest
chunk offset range on disk.

fixes: [ch1531]
fixes: [ch1551]
fixes: [ch1602]

Signed-off-by: Noah Watkins <noah@vectorized.io>