s/spill_key_index: evicting indices using flat map load factor
In spill key index when estimating memory usage we accounted only size
of the keys. Real memory usage of underlying flat map depends on both
value and key size as well as the underlying type capacity. Added
eviction of memory based on the load factor and the actual capacity (not
the size) of indices map.

Signed-off-by: Michal Maslanka <michal@vectorized.io>