s/spill_key_index: made max key length constant
In order to make sure that the truncation of the same key lead to equal
results, we have to always truncate key to the same length.
In the solution before the current fix we truncated the key to variable length
basing on the length of offset and delta varint representation. This
lead to the situation where the same key was treated as two unique keys
as it had the same length.

Example:
original key: 0123456789

truncate to 5 digits: 01234
truncate to 4 digits: 0123

"01234" != "0123"

Signed-off-by: Michal Maslanka <michal@vectorized.io>