rpk/container: Add command to start a cluster
Add a command to start a local cluster with Docker.
It creates everything: The volumes, network and containers.
Requires a redpanda:latest image to be available.

In Linux, the nodes are available (i.e. for `rpk api *`) through
the container's IP + its exposed ports.

In MacOS, the containers can be reached via the host's IP + the
containers' exposed ports.