s/disk_log_impl: preventing eviction of an active segment
Added code preventing eviction of an active segment. When active segment is evicted
it may lead to offset tracker corruption and in the end to data loss.

Currently we to check if there are segments to evict we use committed
instead of base offset. We only allow eviction if of a segment if
calculated max offset greater or equal to segment committed_offset.
Hence we only delete those segments which are designated to be evicted
in total.

Fixes: [ch1663]

Signed-off-by: Michal Maslanka <michal@vectorized.io>