storage: incrementally clean-up batch cache index
When low-memory conditions trigger batch cache reclaims batches were
reclaimed, but the per-segment indicies that map offsets to batches were
not pruned. When we disabled cache-cleaning-on-segment-roll this
resulted in memory pressure due to increased amount of orphaned index
entries.

This patch adds a reference to batch cache entries that points back to
the owning index, allowing the index entry to be removed during reclaim.
Since removing entries from a abseil btree map invalidates iterators, we
also introduce a index-level lock that allows the reclaimer to free
batch memory but defer the actual removal of the index entry until after
the index has been unlocked.

You may notice that redundancy now between the weak pointer feature for
invalidating index entries and the new index back reference. This
redundancy is real and will be removed on the next iteration on the
batch cache design.

Signed-off-by: Noah Watkins <noah@vectorized.io>