storage: fix use-after-free shutdown race in seg appender
prior to the change introduced by this patch the following produced a
race which leads to a background operation referencing a deleted segment
appender:

  {
    segment_appender a;
    a.append(chunk).get();
    a.close().get();
  }

Assume a brand a new segment appender, and one append operation for a
full chunk. The append will set bytes_flush_pending = 16K, and dispatch
a background write operation which decrements bytes_flush_pending by the
I/O size--in this case 16K--causing it go to zero, and then dispatches a
background write operation.

When close is called it first calls flush which returns immediately if
bytes_flush_pending is zero. However, flush is also what introduces a
clean-up barrier by waiting on all semaphore units. Without this barrier
close will immediately return and the segment appender will be destroyed
with an active background write operation holding references to the
appender.

This is address by:

The short circuit optimization is removed in flush. This trivially fixes
the race by always acquiring the semaphore, but also fixes a potential
data loss event: flush solves the problem that counting bytes can't
address: we don't know what has reached disk. Instead flush when
required and rely on it to be efficient when it really has nothing to
do.

Note that there is a related optimization that may be desired which is
to avoid flushing _iff_ no bytes have arrived since the last flush. But
that doesn't appear to be the optimization that bytes_pending was
targeting.

Signed-off-by: Noah Watkins <noah@vectorized.io>