r/consensus: added max consumable offset
Added max_consumable_offset field to consensus to make sure that we have
a sound representation of an offset that may be read by consumers.

Max consumable offset is an offset that is safe to be fetched by the
consumers. This is similar to Kafka's HighWatermark. Max consumable
offset is depends on the consistency level of messages replicated by the
consensus instance and state of the log. Max consumable offset similarly
to HighWaterMark is monotonic and can never move backward. Max
consumable offset is always greater than log start offset and smaller
than log dirty offset.

Max consumable offset is updated in two scenarios

- committed offset is updated (consistency_level=quorum)
- when batch with was appended to the leader log with relaxed
  consistency

We always update max_consumable offset with std::max(prev,
possible_value) to guarantee its monotonicity.

Signed-off-by: Michal Maslanka <michal@vectorized.io>