Makes a prevoting request do not suppress a voting request
After a follower casts a vote it ignores new voting requests for
a while. Since we use a vote request as prevote it may cause a
prevote request to suppress a follow up vote.

This change updates a node not to ignore a vote request in case it
comes from a voted for node thus fixing he prevote blocking vote
problem.