rpc/connection_cache: made shard assignment depend on current node_id
Added current node_id shift to outgoing connections shard assignment.
In order to make sure that connections to the same node created on
different nodes are assigned to different shards added node_id based
offset to connections shard assignment. This approach is required to
make use of deterministic incoming connection to shard assignment on the
target node. If there would be no shift all connections to given node
would be handled on the same set of shards.

Signed-off-by: Michal Maslanka <michal@vectorized.io>