rpk/container: Fix local image lookup
The images list is currently being filtered by comparing the
image ID, which is a SHA, to the image name
'vectorized/redpanda:latest', so the image wasn't ever found.

Search for an image using the 'reference' filter, not comparing
the image ID. With the filter, ImageList will only pick up
images called 'vectorized/redpanda:latest' (of which there can
only be 1).

https://docs.docker.com/engine/reference/commandline/images/#filter-images-by-reference