Fixes Raft nodes not to erase their votes on leader step down
do_step_down function used to erase voted_for information on every
replicated batch. This behaviour is prone to the split brain problem.

The patch updates do_step_down not to erase the votes and updates
consensus to erase the votes only when they become obsolete (after
the term increases)