storage: expose stable offset from segment appender
Adds an upcall to the segment appender which can be used to receive
notifications of the last readable offset on disk. Since many background
writes may be inflight concurrently, a small rolling history of the
active writes must be maintained in order to ack offsets and also avoid
holes in the offset range.

Signed-off-by: Noah Watkins <noah@vectorized.io>