storage: track the active segment stable offset
The segment offset tracker now contains a new stable offset which is the
largest offset that has been written to disk (but may not be flushed).
Ths is useful for readers that want to access batches that have not yet
been committed, but are readable from disk.

The segment receives notfication from the segment appender when the
readable range is extended, and a small history of
batch-to-physical-offset is maintained in the segment and used to
determine the largest logical offset that is visible on disk.

The history maintained is a rolling history which should never be larger
than about one chunk worth of batches, but may temporarily grow larger
if scheduling or slow disk io causes a delay in completing background
writes within the segment appender.

Signed-off-by: Noah Watkins <noah@vectorized.io>