storage: log reader can read up to dirty offset
Adjusts the bounds and stopping conditions of the log reader to expose
more data to the reader. Previously the disk could not be read past the
committed offset, and only ~32KB of data past the committed offset could
be read from the batch cache because committed offset stopping condition
was checked after the cache was read.

This patch changes the stopping condition to use the dirty offset as the
upper bound on visible offsets. This allows all batches up to the dirty
offset to be consumed from the batch cache without stopping the reader.

Finally, when a cache miss occurs it is important to apply an upper
bound limit when reading from disk because in general the dirty offset
has not yet been written and is visible from disk. Thus before reading
from disk we ensure that the range does not pass the segment's stable
offset bound.

Signed-off-by: Noah Watkins <noah@vectorized.io>