Fix race condition in rpc test
The race condition was triggered by the destruction of tmp dir
which removes files in d-tor. File removal triggers a callback
which depends on 'updated'. But at the point when 'tmp' is being
destroyed the 'updated' is already destroyed (because of deterministic
destruction rule). To fix this we have to postprone the destruction of
the 'updated' var.