gobekli: adds log replay to check that we don't lose old events
Consistency validation works only with key/value interface but
redpanda and kafka is a log so in order to check consistency we
treat it as a tx log and fetch and apply the txes to the in-memory
cache. We fetch only the difference between the last cache update
and the offset of the currently inserted txes so we may not notice
if we start losing the head of the log.

This commit adds a background replay of the all log to check that
there is no data loss.