coproc: Pass shared_ptrs to topic_state instead of references
- topic_state instances should still exist when they are removed from the cache
in the event there is an ongoing request with respect to the owning ntp. For
example a deletion occurs on an ntp, while a future is executing a push and
waiting on a reply from the copro engine. We want to still finish handling the
request and commit any extraneous state to the topic_state instance.