coproc: Defer mutations to internal state to occur within run loop
- Previously the add and remove source methods were in no way synchronized with
the run loop, which meant that while the run loop was attempting to do something
with a particular ntp, at the same time it may have been deregistered by
remove_source.

- This change makes that no longer possible by deferring these mutations to the
'_source' cache (of registered ntps) to occur within lock step in the copro run
loop. So right before route_ntp() is called, all pending registrations and
deregistrations are processed, guaranteeing that during the processing of the
run loop '_sources' will contain the keys that were registered during the
time at the start of its latest iteration.

- fixes #160