r/consensus: added timeout to semaphore wait in vote request handler
When node is unable to cast vote because it can not acquire raft state
mutex we should return after a timeout. Not using a timeout in the
request may lead to the situtation where requests which were timed out
on the client are processed by the server which may result in latency
build up.

Signed-off-by: Michal Maslanka <michal@vectorized.io>