r/consensus: do not ignore reordered request that move follower forward
Made filtering reordered append entries reply less strict. i.e. if
the response dirty offset is greater or equal to the leader offset we
can safely accept the response as it will not trigger the recovery.

In our raft implementation reordered heartbeats are first class
citizens. We decide to use this approach to eliminate locking in
heartbeat manager. Requests reordering is tolerated by raft protocol but
may lead to degraded performance. In order to prevent the performance
degradation we associate requests send to followers with sequence number
and filter responses using the number stored in follower state. Sometime
the reordered request may contain information that is correct and should
be accepted by the leader.

Signed-off-by: Michal Maslanka <michal@vectorized.io>