http: fix iobuf_body implementation
The old implementation assumed that http parser from boost.beast
uses buffer that we pass to it's put method. The buffers that it
passes to iobuf_body::value_type::append are assumed to be originating
from that original buffer provided by user's code. But contrary to
that, in some corner cases it can pass buffer produced from stack
allocated array.

Updated implementation of iobuf_body doesn't have this assumption
and safe to use with http parser from boost.beast library. It
performs its intended optimization if the buffer that was passed
to iobuf_body::value_type::append is originating from the original
source buffer. But if the buffer is stack allocated it resorts to
copying.