s/segment_utils: marking segment as compacted after compaction
Before we marked segment as compacted only after verifying compaction
index state. The verification required compaction index to be read and
checksummed. Modified the behavior so that we mark segment as self
compacted after successful compaction. This way we do not have to read
compaction index multiple times as we skip verification of correctly
compacted index.

Signed-off-by: Michal Maslanka <michal@vectorized.io>