r/recovery_stm: we have to stop recovery stm when node is not a leader
Fixed recovery stopping recovery stm when a node is not a leader.

It may happen that the recovery state machine will start when the node
is not longer a leader. It may lead to the situation when non leader
node will be trying to send append entry requests. This is incorrect and
may lead to protocol not being able to make progress.

The need for checking both the term and the leadership comes from the
fact that we may be a leader in two different terms.

Signed-off-by: Michal Maslanka <michal@vectorized.io>