gobekli: decouples kafkakv and redpanda deployment
Kafkakv is a proxy between Kafka and the key/value interface
(consistency testing works with key/value). When we collocate
kafkakv and redpanda on the same nodes we miss some testing
scenarios e.g. a case when kafkakv sees an isolated and an
active redpanda nodes at the same time.

This commit decouples kafkakv and redpanda deployment and makes
it possible to test the mentioned scenario.