rpk/config: Fix Write tests
TestWrite declared a `const path` which was the filepath to which
the `existingConf` field contents were written to. However,
`Write` itself writes the config to the given `Config`'s
`ConfigFile` path, which is /etc/redpanda/redpanda.yaml by default.
That means that `Write` was writing to a different place than the
one that was checked afterwards.