r/consensus: suppressing heartbeats while in replicate stm
When we are replicating data with acks=-1 we have to skip sending
heartbeats to the followers. The heartbeat might be send before an RPC
containing data already appended to this.

Example:

Initial state:
  -> leader dirty log offset = 10
  -> follower is up to date with the leader

1. leader replicate is called with batch containing 5 records
2. leader appends batch to local log (last log offset = 15)
3. heartbeat is created with last log offset = 15
4. heartbeat is sent to the follower
5. append_entries RPC is sent to the follower

follower receives first request with offset=15 which cause it to return
error in append entries reply what it turn triggers recovery on the
leader node.

Signed-off-by: Michal Maslanka <michal@vectorized.io>