rpk/container: Don't rely on volumes
Redpanda runs under the 'redpanda' user in the containers.
When `rpk container start` was run, it mounted the configuration
dir as a volume, messing up the permissions of the volume's target
path (/etc/redpanda)inside the container, which made rpk fail when
it tried to write the config to /etc/redpanda/redpanda.yaml.

Use the --seeds, --kafka-addr, --rpc-addr, --advertise-kafka-addr,
--advertise-rpc-addr & --node-id flags to remove the need for a
config file.