Add file_stm implementation
file_stm is a wrapper around ss::file that allows to store opened
files in the cache (per-shard). This allows us to evict files if
the cache is getting full.

file_stm is implemented as a state machine that can be in opened,
closed, or evicted state. Any attempt to access evicted object
will result in reopening of the file. Also, if any active reader
is present, file_stm object won't be evicted until the reader is
done. The scenario when the file has active reader and at the same
time is a candidate for eviction is very unlikeley, but it handled
anyway.

Also, file_stm can be in a pre_open state. Such file will be opened
on first access.