kafka: add tests for scram algorithm
The test cases are predominently the same scram message tests found in
upstream kafka. This patch does not contain tests for a full exchange
because our code base does not implement the client side of the scram
algorithm. Those tests will be implicit with integration tests when the
final SASL framework with SCRAM support is added.

Signed-off-by: Noah Watkins <noah@vectorized.io>