kafka: add noop incremental alter configs request
Incremental alter configs is similar to the non-incremental alter
configs request but does require the entire configuration to be
specified in one request.

In neither case do we currently support runtime changes to
configurations.

This patch adds a noop implementation for incremental alter configs
request that returns an error indicating the operation is not supported,
but otherwise the API appears to exist from the client's perspective.

Signed-off-by: Noah Watkins <noah@vectorized.io>