coproc/js: change getLastFragment on IOBuf
before it function just find the last fragment on the list in order
to append information, now it finds the first fragment with free
space.

add getNextFragmentOrAppend this function tries to find
the next fragment for given one, if there isn't next fragment it append
the new one and return it.

add clean function for reset state on IOBuf.