r/recovery_stm: fixed triggering read when there are no data
Incorrect check was causing recovery to stop as there were no data to
read for the follower. Fixed this issue by only triggering read for
recovery when there are actual data to read for the follower.

We want leader log dirty offset to be always > because
if it is = it means we are caught up with the follower.

Incorrect handling of recovery caused significant increase in ACKS=1
e2e latency.

Signed-off-by: Michal Maslanka <michal@vectorized.io>