kafka: break circular dep with server and requests
The kafka server depends on the group membership service which in turn
depends on request / response types associated with group membership
apis. However, the headers for these apis contained references to the
request context which we'd like to be able to be used to provide access
to the server.

This patch moves references to the request context out of the header,
breaking that dependency and allowing the request context to access the
server type.

Signed-off-by: Noah Watkins <noah@vectorized.io>