r/consensus: fixed updating follower state when its data were deleted
When follower was up to date with the leader and its data were deleted
it may happen that recovery STM will never be trigger as the follower
state is not updated. Added updating follower state (its dirty and
committed offset) when its dirty index moved backward.

NOTE: this problem caused `test_raft_rpunit` to fail sometimes.

Signed-off-by: Michal Maslanka <michal@vectorized.io>