rpk/status: Run getMetrics sychronously
Only the system resources usage metrics are sent when --send is passed.
`rpk status --send` is rarely run by a human, and it's often run by the
redpanda-status.service (for deb & rpm installations) or the cron timer
in the Docker image. These processes don't care about `rpk status`'s
output, so it makes sense to retrieve the metrics synchronously (this
patch), check if --send was passed, and if so, just send the data and
return (next patches).