ci: Add consistency test workflow
The workflow is initiated through a slash command in PRs (`/chaos-test`).
Packages have to be available for the HEAD of the PR
(which will happen automatically and be visible in the PR).

Results are published to an AWS S3 bucket and a comment with test result
along with their location is left in the PR.