kafka: use std::regex with gcc to avoid memory bug
Likely an issue with gcc, memory usage is unacceptably high using c++20 with
these regular expressions and the ctre library. Since this is not a problem
with clang and our release builds use clang, we fall back to slow std::regex
when using gcc. The following ticket is tracking the issue, and the fallback
can be removed once the issue is resolved:

https://github.com/hanickadot/compile-time-regular-expressions/issues/155

Signed-off-by: Noah Watkins <noah@vectorized.io>