r/consensus: do not promote to voter if node is not in configuration
Currently if node is no longer in configuration it may be promoted to
voter as `raft::configuration::is_voter()` return `false` when node is
not present in configuration - this is correct as node is not a voter.
Added explicit check to make sure that node is part of current
configuration before promoting it from learner to voter.

Signed-off-by: Michal Maslanka <michal@vectorized.io>