rptest/redpanda: tweak find_binary method
Assume a global `rp_install_path_root` variable is given that points to
the installation folders. In addition, check if the underlying platform
is docker-compose (via a global `platform` variable), in which case the
configuration file is written; otherwise, it assumes redpanda is
configured externally (e.g. via ansible).

The build root directory is also expected to be defined via the global
variable `v_build_dir`. This can be used by ducktape tests to find
binaries for other languages (java, javascript, etc.).