http: Add HTTP prefetch
Naturally, http header is not available until enough data is received
from the server. Existing http::client::response_stream interface
allows to check if the header is available but it doesn't allow
to make sure that this is the case. This makes api client interface
inconvenient because the user of such api have to read from the
response_stream, wait until the header is available and then decide if
the result is OK or an error was triggered.

This commit adds 'prefetch_headers' method that reads data from the
socket until the header becomes available for fetching. It's supposed
to be used for error handling in situations when we need to check
the response status first and then proceed with normal operation.

The 'prefetch_headers' implementation is based on existing 'recv_some'
method. It receives data until the header is ready. If any payload bytes
was received with the header they're stored inside the 'response_stream'
object and returned with the first 'recv_some' call.

The test suite checks different framings with and without
'prefetch_headers' being called.