r/consensus: adds conditional replicate
For the idempotent producer feature we need to be sure that when a
broker (a leader) replicates a message it has seen all the previous
messages.

Every time a leader is elected it receives a unique term. If we use
a condition on the current leader's term it guaratees that when an
re-election happens since the last messages catch up the replicate
request doesn't pass.