k/tests: fixed failing topic recreation tests
The topic recreate test had a flaw. The test incorrectly assumed that
once shard is is present it will be present all the time. Currently
during the bootstrap process if topics are created/removed which may
cause the state to be transient. There may be a point in time during
bootstrap, after topic was deleted and then back created, when shard_id
is present and then it become absent again.

Signed-off-by: Michal Maslanka <michal@vectorized.io>