storage: add adjacent segment compaction
A simple implementation of compacting adjacent segments. The basic
strategy is to find two segments whose combined size does not exceed
a configurable max threshold size. This is scheduling, and results in a
pair of segments.

This pair of segments is then physically combined to replace the first
segment. The resulting combined segment is then self-compacted as normal
and then redundant second segment is then permenantely removed.

Signed-off-by: Noah Watkins <noah@vectorized.io>