Update make_iobuf_output_stream
The function is not used anywhere in the code so far. It accepts
iobuf as an input and returns a stream. This stream can be used to
fill the iobuf with data. The problem is that the iobuf is passed
by value and there is no way to retreive it back. This commit chanes
function parameter to iobuf& so the iobuf don't need to be moved and
can be used afterwards.