k/tests: fixed assertion in test_recreate_topic
Fixed redundant assertion in topic recreation test.

In redpanda we optimize topic leadership querying. When topic is created
we assume that leader is elected immediately. There may be a short
period however during which the leader election happens and it updates
the cached leader id to (-1) which means that election is in progress.
Checking metadata twice sometimes caused us to query for leadership when
election was taking place which caused test to fail.

Signed-off-by: Michal Maslanka <michal@vectorized.io>