model: generating_record_batch_reader to take batches of batches
- The future returned by the function passed to the reader will produce batches
of batches instead of a single record_batch. This gives more control to users
who may want to produce only one record_batch or more than one

- Previously the future following the call to the generator method, checked to see if
the return value, wrapped in an ss::optimized_optional, was true/false. If true, this
signified the end of stream. Now to singal end of stream the method checks for an
empty batch. Theres no need for those typedefines or optionals here.