operator/controller: Inject node-specific configuration to SS pods
The StatefulSet is provided with a single configMap containing common
configuration across the Pods. At init time the configMap is copied
as a Redpanda configuration file to the Pod. Since each the conf files
across nodes must be different to account for node id and advertised
address, we call rpk at the init time to adjust the redpanda conf file
to reflect the node details.