arch: Add manifest implementation
Manifest is an object needed to track objects in S3 and map them
to and from local log segments. One manifest is created per ntp.
It also takes revision into account, so when partition gets deleted
and created it will have new manifest. Manifest contains information
about the partition and about every segment located in S3. It can be
used to map local log segments to S3 locations and back. It can also
be (de)serialized to json.

There is a diff fucntion implementaiton that is supposed to be used
to find candidates for upload or removal. The archival service
creates one manifest for local data and one for remote (stored in S3)
and uses this funciton to find the upload/remove candidates.
Manifest is supposed to be stored in S3 bucket.