model: support multiple kafka listeners per broker
This patch introduces a wire protocol and storage format breaking
change.

Extends types to support associating multiple named listeners per
broker. The main type introduced is struct broker_endpoint which is a
(name, address) pair.

model::broker is extended to store a set of broker endpoints. existing
tests have been built assuming a single listener, and these tests are
updated to take the first listener in the set which should provide
compatability until the tests need to be generalized.

Signed-off-by: Noah Watkins <noah@vectorized.io>