r/tests: fixed assertion in conditional replicate tests
The conditional replicate tests were relaying on an assumption that the
leadership will be stable during the test while there are no
failures injected. However, the small election timeouts that are used
in the tests may lead to the situation when leader election will happen
more than once during the test. (This is somewhat expected as it tests
more scenarios). Added validation that checks if term really didn't
change during the test run.

Signed-off-by: Michal Maslanka <michal@vectorized.io>