coproc/js: convert findByGlobalId from arrow fn to class fn
The arrow function doesn't bind to class scope, for that reason
this function isn't possible to mock, when we convert it to normal
class fn, when this class is instanced this function is added to
Class.Prototype and in this case this function is part of class context.
Also it improves performance.